Refactoring deabstraction as a helper function.
