+++ lib/SILOptimizer/PassManager/PassPipeline.cpp
@@ -634,7 +634,6 @@ SILPassPipelinePlan SILPassPipelinePlan::getOnonePassPipeline() {
 SILPassPipelinePlan SILPassPipelinePlan::getTFPartitionPassPipeline() {
   SILPassPipelinePlan P;
   P.startPipeline("TensorFlow Partitioning");
-  P.addTFDeabstraction();
   P.addTFPartition();
   return P;
 }
